[FEATURE] User Configuration
Add UserConfig model classes. Add ConfigHandler class.
Add method to ConversionHelper class that returns an enumeration of items from some lines that start with one of a number of prefixes. Refactor where explicit splitting of lines was done.

[TEST][FEATURE] initialise ConfigHandler in Converter

Add initialisation of ConfigHandler in the Converter.
Add tests for ConfigHandler.

[TEST] Added tests for configuration

Refactored converter tests into separate classes by feature being tested. Added ConverterConfigTests class.
Moved ValidateOptions into new OptionsHelper class.
Added ReflectionHelper class.
