+++ Crowswood.CsvConverter/UserConfig/GlobalConfig.cs
@@ -1,0 +1,7 @@
+﻿namespace Crowswood.CsvConverter.UserConfig
+{
+    internal sealed class GlobalConfig : BaseConfig
+    {
+        public GlobalConfig(string name, string value) : base(name, value) { }
+    }
+}
